requirements.txt
boto3
colorama
jinja2
pyfiglet
termcolor
PyYAML